requirements.txt
asgiref==3.7.2
Babel==2.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-tailwind==1.0.3
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-allauth-ui==0.2.2
django-browser-reload==1.12.1
django-crispy-forms==2.1
django-money==3.4.1
django-phonenumber-field==7.3.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
idna==3.6
install==1.3.5
mypy==1.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
phonenumbers==8.13.32
phonenumberslite==8.13.32
psycopg==3.1.18
psycopg-binary==3.1.18
py-moneyed==3.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
urllib3==2.2.1
